feat(ui, __main__.py, application.py): add main window UI and application code
Add a new main window UI file and the necessary code to display it in the
application. The application code includes a logger setup and the main
function that initializes the application and shows the main window.

feat(main_window.py, app_context.py, app_settings.py): add main window, app context, and app settings modules

Add the main window module to create the main window of the application. Add the app context module to create a context object that holds the PyJapc instance used to communicate with the SPS accelerator. Add the app settings module to hold the application settings.

feat(app_settings.py): add AppSettings class to wrap QSettings and provide easy access to some data that can be saved to the user's computer. Add methods to save and retrieve window state and last used directory in a file browser.